requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\GovnoWare\Desktop\TelePromo\TelePromo.py: 1
Telethon == 1.16.0

# C:\Users\GovnoWare\Desktop\TelePromo\TelePromo.py: 7
discord.py == 1.4.1
